MC-85494 Изменен тип данных с int на double для полей weight и volume в сущности Bundle

+++ src/main/java/ru/moysklad/remap_1_2/entities/products/Bundle.java
@@ -51,12 +51,12 @@ public class Bundle extends AbstractProduct implements IEntityWithAttributes<Att
     /**
      * Вес
      */
-    private Integer weight;
+    private Double weight;
 
     /**
      * Объём
      */
-    private Integer volume;
+    private Double volume;
 
     /**
      * Отметка о том, добавлен ли Комплект в архив
